% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_tourist.R
\name{calculate_distance_tourist}
\alias{calculate_distance_tourist}
\title{Construction of distance function for tourist data}
\usage{
calculate_distance_tourist(x, y, dist_file)
}
\arguments{
\item{x}{observation 1}

\item{y}{observation 2}

\item{dist_file}{path to overview file about variables}
}
\value{
distance between two observations
}
\description{
Function to define a valid distance function for the tourist dataset
}
