% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{calculate_memb_score}
\alias{calculate_memb_score}
\title{Calculate membership score of one observation for each medoid}
\usage{
calculate_memb_score(dist_med, m)
}
\arguments{
\item{dist_med}{Vector of distances to medoids}

\item{m}{Fuzziness exponent, which has to be a numeric of minimum 1. Defaults
to 2.}
}
\value{
List with membership scores for one observation
}
\description{
Function to calculate a membership score for one observation
for each medoid based on the distance of this observation to all medoids
}
