% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_silhouettes.R
\name{calculate_silhouettes}
\alias{calculate_silhouettes}
\title{Calculatilon of silhouettes}
\usage{
calculate_silhouettes(object)
}
\arguments{
\item{object}{an object of class claraclust}
}
\description{
Function to calculate silhouettes for each clustering observation
}
