% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_bar_boxplot}
\alias{clara_bar_boxplot}
\title{Plot function barplot or boxplot}
\usage{
clara_bar_boxplot(x, data, variable, group_by = NULL, na.omit = FALSE)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are
already filtered by threshold (fuzzy))}

\item{variable}{name of variable to plot}

\item{group_by}{optional grouping variable}
}
\value{
barplot or boxplot
}
\description{
Function to plot a barplot or a boxplot
}
