% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_parallel}
\alias{clara_parallel}
\title{Plot function parallel coordinate plot}
\usage{
clara_parallel(x, data, membership_threshold = 0, seed = 42)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{Prepared data.frame (contains cluster variable, observations are
already filtered by threshold (fuzzy))}

\item{membership_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}

\item{seed}{random number seed}
}
\value{
parallel coordinate plot
}
\description{
Function to plot a parallel coordinate plot
}
