% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_silhouette}
\alias{clara_silhouette}
\title{Plot function silhouette}
\usage{
clara_silhouette(
  x,
  data,
  metric = "euclidean",
  silhouette_subsample = FALSE,
  scale_sil = TRUE,
  rel_obs = NULL
)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{Prepared data.frame (contains cluster variable, observations are
already filtered by threshold (fuzzy))}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{silhouette_subsample}{Use the subsample from 'x' for silhouette
plot instead of all samples? Defaults to FALSE.}

\item{scale_sil}{Scale numeric variables for silhouette plot? Defaults to
TRUE. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{rel_obs}{Optional names of observations > threshold.}
}
\value{
silhouette plot
}
\description{
Function to plot a scatterplot
}
