% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_wordcloud}
\alias{clara_wordcloud}
\title{Plot function wordcloud}
\usage{
clara_wordcloud(x, data, variable, seed = 42)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{variable}{name of variable to plot}

\item{seed}{optional seed}
}
\value{
wordcloud plot
}
\description{
Function to plot a wordcloud
}
