% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_clarans.R
\name{clustering_local}
\alias{clustering_local}
\title{Perform a local iteration of CLARANS clustering}
\usage{
clustering_local(
  data,
  sample_local,
  clusters = 5,
  metric = "euclidean",
  max_neighbors = 100,
  type = "fixed",
  m = 2,
  verbose = 1,
  verbose_toLogFile = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{sample_local}{list containing information on pairs of medoids and
non-medoids tested for swapping as well as starting medoids for the
algorithm}

\item{clusters}{Number of clusters. Defaults to 5.}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Will be passed as argument
\code{method} to \code{\link[proxy]{dist}}, so check \code{?proxy::dist} for
full details.}

\item{max_neighbors}{Maximum number of randomized medoid searches with each
cluster (only if \code{algorithm = "clarans"})}

\item{type}{One of \code{c("fixed","fuzzy")}, specifying the type of
clustering to be performed.}

\item{m}{Fuzziness exponent (only for \code{type = "fuzzy"}), which has to be
a numeric of minimum 1. Defaults to 2.}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more
detailed messages. Defaults to 1.}

\item{verbose_toLogFile}{If TRUE, the diagnostic messages are printed to
a log file \code{clustering_progress.log}. Defaults to FALSE.}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
Clustering solution for data sample
}
\description{
Function to perform a local iteration of the clarans clustering algorithm in
a fixed or fuzzy way. The function can either be performed using a common
dissimilarity metric or a self-defined distance function.
}
