% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{clustering_sample}
\alias{clustering_sample}
\title{Perform clustering algorithm on a data sample}
\usage{
clustering_sample(
  data,
  sample_ids,
  clusters = 5,
  metric = "euclidean",
  sample_size = NULL,
  type = "fixed",
  m = 2,
  verbose = 1,
  verbose_toLogFile = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{sample_ids}{ids for the sample's observations}

\item{clusters}{number of clusters}

\item{metric}{predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{sample_size}{number of observations belonging to a sample}

\item{verbose_toLogFile}{If TRUE, the diagnostic messages are printed to
a log file \code{clustering_progress.log}. Defaults to FALSE.}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
clustering solution for data sample
}
\description{
Function to perform clara clustering algorithm in a fixed or fuzzy way.
The function can either be performed using a common dissimilarity metric or
a self-defined distance function.
}
