% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variable_weights.R
\name{compute_variable_weights}
\alias{compute_variable_weights}
\title{Computation of individual weights for each variable of a distance file}
\usage{
compute_variable_weights(dist_file)
}
\arguments{
\item{dist_file}{data.frame containing information about variables
supposed to be used for clustering}
}
\value{
dist_file with new column "weight_total"
}
\description{
Function to compute individual weights for variables given by a distance
file
}
