% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyclara.R
\name{fuzzyclara}
\alias{fuzzyclara}
\title{Perform clustering}
\usage{
fuzzyclara(
  data,
  clusters = 5,
  metric = "euclidean",
  algorithm = "clara",
  samples = 10,
  sample_size = NULL,
  max_neighbors = 100,
  num_local = 10,
  type = "fixed",
  cores = 1,
  seed = 1234,
  m = 2,
  verbose = 1,
  scale = TRUE,
  build = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{clusters}{Number of clusters}

\item{metric}{Predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{algorithm}{One of \code{c("clara","clarans")}}

\item{samples}{Number of subsamples (only for clara clustering)}

\item{sample_size}{Number of observations belonging to a sample. If NULL
(default), the minimum of \code{nrow(data)} and \code{40 + clusters * 2} is
used as sample size. (only for clara clustering)}

\item{max_neighbors}{Maximum number of randomized medoid searches with each
cluster (only for clarans clustering)}

\item{num_local}{Number of clustering iterations (only for clarans
clustering)}

\item{type}{One of \code{c("fixed","fuzzy")}, specifying the type of
clustering to be performed.}

\item{cores}{Numbers of cores for computation (cores > 1 implies
multithreading)}

\item{seed}{Random number seed}

\item{m}{Fuzziness exponent (only for \code{type = "fuzzy"})}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more detailed
messages. Defaults to 1.}

\item{scale}{Scale numeric variables before distance matrix calculation?
Default TRUE}

\item{build}{Additional build algorithm to choose initial medoids (only
relevant for type = "fuzzy". Default FALSE.)}

\item{...}{Additional arguments passed to the main clustering algorithm and
to proxy::dist for the calculation of the distance matrix
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
Object of class claraclust
}
\description{
Function to perform a cluster analysis in a fixed or fuzzy way. The
function can either be performed using a common dissimilarity metric or
a self-defined distance function.
}
\details{
If the clustering is run on mulitple cores, the verbose messages are printed
in a file \code{clustering_progress.log} (if \code{verbose > 0}).
}
