% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{perform_sample_clustering}
\alias{perform_sample_clustering}
\title{Perform pam or vegclust clustering on a data sample}
\usage{
perform_sample_clustering(
  dist,
  clusters,
  type,
  names,
  m = 2,
  build = FALSE,
  ...
)
}
\arguments{
\item{dist}{Dissimilarity matrix}

\item{clusters}{Number of clusters}

\item{type}{Fixed or fuzzy clustering}

\item{names}{Vector of names for observations}

\item{m}{Fuzziness exponent (only for type = fuzzy)}

\item{build}{Additional build algorithm to choose initial medoids (only
relevant for type = "fuzzy". Default FALSE.)}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
List with information on cluster results (medoid and cluster
assignment)
}
\description{
Function to perform pam in a fixed or fuzzy way on a data sample
}
