% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{plot.fuzzyclara}
\alias{plot.fuzzyclara}
\title{Visualization of clustering solution by variables}
\usage{
\method{plot}{fuzzyclara}(x, data, type = NULL, variable = NULL, na.omit = FALSE, ...)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{data.frame or matrix used for clustering}

\item{type, variable}{Type of plot. One of \code{c("barplot","boxplot","wordclouds",
"silhouette","pca","scatterplot")}. Defaults to NULL, which either plots
a barplot or a boxplot, depending on the class of \code{variable}.}

\item{na.omit}{Should missing values be excluded for plotting? Defaults to
FALSE.}

\item{...}{Further arguments for internal plot functions.}
}
\value{
Clustering plot
}
\description{
Function to provide graphical visualization of distribution
}
