% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_cluster_numbers.R
\name{plot_cluster_numbers}
\alias{plot_cluster_numbers}
\title{Visualization of the selection criterion for different cluster numbers}
\usage{
plot_cluster_numbers(
  data,
  clusters_range = 2:5,
  metric = "euclidean",
  samples = 10,
  sample_size = NULL,
  type = "fixed",
  cores = 1,
  seed = 1234,
  m = 2,
  verbose = 1,
  return_results = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{clusters_range}{range for numbers of clusters}

\item{metric}{predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{samples}{number of subsamples}

\item{sample_size}{number of observations belonging to a sample. If NULL
(default), the minimum of \code{nrow(data)} and \code{40 + clusters * 2} is
used as sample size.}

\item{type}{fixed or fuzzy clustering}

\item{cores}{numbers of cores for computation (cores > 1 implies
multithreading)}

\item{seed}{random number seed}

\item{m}{fuzziness exponent (only for type = "fuzzy")}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more detailed
messages. Defaults to 1.}

\item{return_results}{return clustering results (claraclust objects) as a list}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
object of class claraclust
}
\description{
Function to provide graphical visualization for selecting the optimal number of clusters
}
