% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fuzzyclara.R
\name{predict.fuzzyclara}
\alias{predict.fuzzyclara}
\title{Prediction of cluster assignments}
\usage{
\method{predict}{fuzzyclara}(object, newdata, return_distMatrix...)
}
\arguments{
\item{object}{an object of class claraclust}

\item{newdata}{data.frame to look variables with which to predict}

\item{return_distMatrix}{Should the distanceds to the cluster medoids be
returned?}

\item{...}{further arguments for predict functions}
}
\value{
clustering plot
}
\description{
Function to predict cluster assignments
}
