% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_tourist.R
\name{prepare_variable_names}
\alias{prepare_variable_names}
\title{Construction of distance function for tourist data}
\usage{
prepare_variable_names(dist_file)
}
\arguments{
\item{dist_file}{variable names column of dist_file}
}
\value{
variable names corresponding to the names in the tourist data
}
\description{
Function to adjust the variable names, especially for "nominal mehrfach" version
}
