% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Assign a cluster to each observation of the entire dataset}
\usage{
assign_cluster(
  data,
  metric,
  medoids,
  type = "fixed",
  m = 2,
  return_distMatrix = FALSE,
  return_data_medoids = FALSE
)
}
\arguments{
\item{data}{entire data.frame}

\item{metric}{predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{medoids}{medoids of the obtained clustering solution for the data
sample}

\item{type}{fixed or fuzzy clustering}

\item{m}{fuzziness exponent (only for type = fuzzy)}

\item{return_distMatrix}{Should the distances to the cluster medoids be
returned?}

\item{return_data_medoids}{Should medoid data be returned? (This could be
used for assigning new observations to the clusters.)}
}
\value{
list with information on cluster results (medoid, cluster
assignment, average distance to the closest medoid (weighted
average distance to the closest medoid in case of fuzzy clustering))
}
\description{
Function to assign a cluster to each observation of the entire dataset
by selecting the closest medoid
}
