% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Assign a cluster to each observation of the entire dataset}
\usage{
assign_cluster(
  data,
  metric,
  medoids,
  dist_file,
  type = "fixed",
  m = 2,
  return_distMatrix = FALSE
)
}
\arguments{
\item{data}{entire data.frame}

\item{metric}{predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{medoids}{medoids of the obtained clustering solution for the data
sample}

\item{dist_file}{scheme for TourIST distance calculation}

\item{type}{fixed or fuzzy clustering}

\item{m}{fuzziness exponent (only for type = fuzzy)}
}
\value{
list with information on cluster results (medoid, cluster
assignment, average distance to the closest medoid (weighted
average distance to the closest medoid in case of fuzzy clustering))
}
\description{
Function to assign a cluster to each observation of the entire dataset
by selecting the closest medoid
}
