% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Assign a cluster to each observation of the entire dataset}
\usage{
assign_cluster(
  data,
  metric,
  medoids,
  dist_matrix = NULL,
  type = "fixed",
  m = 2,
  return_distMatrix = FALSE,
  return_data_medoids = FALSE
)
}
\arguments{
\item{data}{Entire data.frame}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Will be passed as argument
\code{method} to \code{\link[proxy]{dist}}, so check \code{?proxy::dist} for
full details.}

\item{medoids}{Medoids of the obtained clustering solution for the data
sample}

\item{type}{One of \code{c("fixed","fuzzy")}, specifying the type of
clustering to be performed.}

\item{m}{Fuzziness exponent (only for \code{type = "fuzzy"}), which has to be
a numeric of minimum 1. Defaults to 2.}

\item{return_distMatrix}{Indicator if the distances to the cluster medoids
should be returned. Defaults to FALSE.}

\item{return_data_medoids}{Indicator if the medoid data should be returned.
Defaults to FALSE.}
}
\value{
List with information on cluster results (medoid, cluster
assignment, average distance to the closest medoid (weighted
average distance to the closest medoid in case of fuzzy clustering))
}
\description{
Function to assign a cluster to each observation of the entire dataset
by selecting the closest medoid
}
