% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_clarans.R, R/clustering_sample.R
\name{calculate_memb_score}
\alias{calculate_memb_score}
\title{Calculate membership score of one observation for each medoid}
\usage{
calculate_memb_score(dist_med, m)

calculate_memb_score(dist_med, m)
}
\arguments{
\item{dist_med}{vector of distances to medoids}

\item{m}{fuzziness exponent (only for type = fuzzy)}
}
\value{
list with membership scores for one observation

list with membership scores for one observation
}
\description{
Function to calculate a membership score for one observation
for each medoid based on the distance of this observation to all medoids

Function to calculate a membership score for one observation
for each medoid based on the distance of this observation to all medoids
}
