% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_claraclust.R
\name{clara_pca}
\alias{clara_pca}
\title{Plot function PCA}
\usage{
clara_pca(x, data, group_by = NULL)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{group_by}{optional grouping variable}
}
\value{
PCA plot
}
\description{
Function to plot PCA results
}
