% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_pca}
\alias{clara_pca}
\title{Plot function PCA}
\usage{
clara_pca(
  x,
  data,
  group_by = NULL,
  plot_all_fuzzy = TRUE,
  confidence_threshold = 0,
  alpha_fuzzy = 0.4,
  focus = FALSE,
  focus_clusters = NULL
)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{data.frame or matrix used for clustering}

\item{group_by}{Optional grouping variable}

\item{plot_all_fuzzy}{For fuzzy clustering and threshold: should observations
below threshold be plotted transparent? Defaults to TRUE.}

\item{confidence_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}

\item{alpha_fuzzy}{Alpha value for observations below threshold, only
relevant for \code{plot_all_fuzzy = TRUE}. Defaults to 0.4.}

\item{focus}{For fuzzy clustering, focus on clusters given by variable
\code{focus_clusters} and plot observations based on probability of belonging
to the respective cluster. Defaults to FALSE.}

\item{focus_clusters}{Optional vector of integers to focus on specific
clusters.}
}
\value{
PCA plot
}
\description{
Function to plot PCA results
}
