% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_scatterplot}
\alias{clara_scatterplot}
\title{Plot function scatterplot}
\usage{
clara_scatterplot(
  x,
  data,
  x_var,
  y_var,
  plot_all_fuzzy = FALSE,
  transparent_obs = NULL,
  alpha_fuzzy = 0.4,
  focus = FALSE,
  focus_clusters = NULL
)
}
\arguments{
\item{x}{An object of class claraclust}

\item{data}{Prepared data.frame (contains cluster variable, observations are
already filtered by threshold (fuzzy))}

\item{x_var, y_var}{Names of x and y variable}

\item{plot_all_fuzzy}{For fuzzy clustering and threshold: should observations
below threshold be plottet transparent? The regression line is only based on
the observations above the threshold. Defaults to FALSE.}

\item{transparent_obs}{data.frame containing observations that are plotted
transparent, only relevant for \code{plot_all_fuzzy = TRUE}.}

\item{alpha_fuzzy}{Alpha value for observations below threshold, only
relevant for \code{plot_all_fuzzy = TRUE}. Defaults to 0.4.}

\item{focus}{For fuzzy clustering, focus on clusters given by variable
\code{focus_clusters} and plot observations based on probability of belonging
to the respective cluster. Defaults to FALSE.}

\item{focus_clusters}{Optional vector of integers to focus on specific
clusters}
}
\value{
scatterplot
}
\description{
Function to plot a scatterplot
}
