% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_claraclust.R
\name{clara_scatterplot}
\alias{clara_scatterplot}
\title{Plot function scatterplot}
\usage{
clara_scatterplot(
  x,
  data,
  x_var,
  y_var,
  transparent_obs = NULL,
  plot_all_fuzzy = FALSE,
  alpha_fuzzy = 0.4,
  focus = FALSE,
  focus_clusters = NULL
)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{x_var}{name of x variable}

\item{y_var}{name of y variable}

\item{transparent_obs}{dataset containing observatiosn that are plottet transparent, only relevant for plot_all_fuzzy = TRUE}

\item{plot_all_fuzzy}{for fuzzy clustering and threshold: should observations below threshold be plottet transparent? The regression line is only based on the observations above the threshold}

\item{alpha_fuzzy}{alpha value for observations below threshold, only relevant for plot_all_fuzzy = TRUE}

\item{focus}{for fuzzy clustering, focus on clusters given by variable focus_clusters and plot observations based on probability of belonging to the respective cluster}

\item{focus_clusters}{vector of integers: on which clusters should be focused? If NULL, all clusters are used}
}
\value{
scatterplot
}
\description{
Function to plot a scatterplot
}
