% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_claraclust.R
\name{clara_scatterplot}
\alias{clara_scatterplot}
\title{Plot function scatterplot}
\usage{
clara_scatterplot(
  x,
  data,
  x_var,
  y_var,
  transparent_obs = NULL,
  plot_all_fuzzy = FALSE
)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{x_var}{name of x variable}

\item{y_var}{name of y variable}
}
\value{
scatterplot
}
\description{
Function to plot a scatterplot
}
