% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_silhouette}
\alias{clara_silhouette}
\title{Plot function silhouette}
\usage{
clara_silhouette(
  x,
  data,
  metric = "euclidean",
  silhouette_subsample = FALSE,
  scale_sil = TRUE,
  silhouette_table = TRUE,
  membership_threshold = 0
)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{data.frame or matrix used for clustering}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{silhouette_subsample}{Use the subsample from 'x' for silhouette
plot instead of all samples? Defaults to FALSE.}

\item{scale_sil}{Scale numeric variables for silhouette plot? Defaults to
TRUE. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{silhouette_table}{Should a table of silhouette results be returned
together with the plot? Defaults to TRUE.}

\item{membership_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}
}
\value{
silhouette plot
}
\description{
Function to plot a scatterplot
}
