% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_silhouette}
\alias{clara_silhouette}
\title{Plot function silhouette}
\usage{
clara_silhouette(
  x,
  data,
  metric = "Euclidean",
  silhouette_subsample = FALSE,
  scale_sil = TRUE,
  rel_obs = NULL
)
}
\arguments{
\item{x}{An object of class claraclust}

\item{data}{Prepared data.frame (contains cluster variable, observations are
already filtered by threshold (fuzzy))}

\item{metric}{Distance metric for silhouette plot. Defaults to
\code{"Euclidean"}. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{silhouette_subsample}{Use the subsample from claraclust for silhouette
plot instead of all samples? Defaults to FALSE.}

\item{scale_sil}{Scale numeric variables for silhouette plot? Defaults to
TRUE. Irrelevant if \code{silhouette_subsample} is TRUE.}

\item{rel_obs}{Names of observations > threshold.}
}
\value{
silhouette plot
}
\description{
Function to plot a scatterplot
}
