% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_claraclust.R
\name{clara_silhouette}
\alias{clara_silhouette}
\title{Plot function silhouette}
\usage{
clara_silhouette(
  x,
  data,
  metric = "Euclidean",
  silhouette_subsample = FALSE,
  scale_sil = TRUE
)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{metric}{distance metric for silhouette plot, default is euclidean. Irrelevant if silhouette_subsample is TRUE}

\item{silhouette_subsample}{use the subsample from claraclust for silhouette plot instead of all samples}

\item{scale_sil}{scale numeric variables for silhouette plot? Default TRUE, irrelevant if silhouette_subsample is TRUE}
}
\value{
silhouette plot
}
\description{
Function to plot a scatterplot
}
