% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{clara_wordcloud}
\alias{clara_wordcloud}
\title{Plot function wordcloud}
\usage{
clara_wordcloud(
  x,
  data,
  variable,
  na.omit = na.omit,
  seed = 42,
  confidence_threshold = 0
)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{Prepared data.frame (contains cluster variable, observations are already filtered by threshold (fuzzy))}

\item{variable}{Name of variable to plot}

\item{na.omit}{Should missing values be excluded for plotting? Defaults to
FALSE.}

\item{seed}{Random number seed. Defaults to 42.}

\item{confidence_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}
}
\value{
wordcloud plot
}
\description{
Function to plot a wordcloud
}
