% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_clarans.R
\name{clustering_clarans}
\alias{clustering_clarans}
\title{Perform CLARANS clustering}
\usage{
clustering_clarans(
  data,
  clusters = 5,
  metric = "euclidean",
  type = "fixed",
  num_local = 5,
  max_neighbors = 100,
  cores = 1,
  seed = 1234,
  m = 2,
  verbose = 1,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{clusters}{Number of clusters. Defaults to 5.}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Will be passed as argument
\code{method} to \code{\link[proxy]{dist}}, so check \code{?proxy::dist} for
full details.}

\item{type}{One of \code{c("fixed","fuzzy")}, specifying the type of
clustering to be performed.}

\item{num_local}{Number of clustering iterations. Defaults to 5.
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}

\item{max_neighbors}{Maximum number of randomized medoid searches with each
cluster. Defaults to 100.}

\item{cores}{Numbers of cores for computation. \code{cores > 1} implies
a parallel call. Defaults to 1.}

\item{seed}{Random number seed. Defaults to 1234.}

\item{m}{Fuzziness exponent (only for \code{type = "fuzzy"}), which has to be
a numeric of minimum 1. Defaults to 2.}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more
detailed messages. Defaults to 1.}

\item{...}{Additional arguments passed to the main clustering algorithm and
to proxy::dist for the calculation of the distance matrix
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
Object of class fuzzyclara
}
\description{
Function to perform a CLARANS clustering in a fixed or fuzzy way.
The function can either be called using a common dissimilarity metric or
a self-defined distance function.
}
\details{
If the clustering is run on mulitple cores, the verbose messages are printed
in a file \code{clustering_progress.log} (if \code{verbose > 0}).
}
\references{
TODO add CLARANS paper
}
