% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_clarans.R
\name{clustering_clarans}
\alias{clustering_clarans}
\title{Perform clarans clustering algorithm}
\usage{
clustering_clarans(
  data,
  clusters = 5,
  metric = "euclidean",
  type = "fixed",
  max_neighbors = 100,
  num_local = num_local,
  cores = 1,
  seed = 1234,
  m = 2,
  verbose = 1,
  ...
)
}
\arguments{
\item{data}{data.frame to be clustered}

\item{clusters}{Number of clusters}

\item{metric}{Predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}

\item{type}{Fixed or fuzzy clustering}

\item{max_neighbors}{Maximum number of randomized medoid searches with each
cluster}

\item{num_local}{Number of clustering iterations}

\item{cores}{Numbers of cores for computation (cores > 1 implies
multithreading)}

\item{seed}{Random number seed}

\item{m}{Fuzziness exponent (only for type = "fuzzy")}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more detailed
messages. Defaults to 1.}

\item{...}{Additional arguments passed to the main clustering algorithm and
to proxy::dist for the calculation of the distance matrix
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
Object of class fuzzyclara
}
\description{
Function to perform clara clustering algorithm in a fixed or fuzzy way.
The function can either be performed using a common dissimilarity metric or
a self-defined distance function.
}
\details{
If the clustering is run on mulitple cores, the verbose messages are printed
in a file \code{clustering_progress.log} (if \code{verbose > 0}).
}
