% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{Compute the dissimilarity matrix for a data sample}
\usage{
compute_distance_matrix(data, metric)
}
\arguments{
\item{data}{sample of data.frame to be clustered}

\item{metric}{predefined dissimilarity metric (euclidean, manhattan) or
self-defined dissimilarity function}
}
\value{
dissimilarity matrix for data sample
}
\description{
Function to compute the dissimilarity matrix based on a specified metric
}
