% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{perform_sample_clustering}
\alias{perform_sample_clustering}
\title{Perform pam or vegclust clustering on a data sample}
\usage{
perform_sample_clustering(dist, clusters, type, names, m, build, ...)
}
\arguments{
\item{dist}{dissimilarity matrix}

\item{clusters}{number of clusters}

\item{type}{fixed or fuzzy clustering}

\item{names}{vector of names for observations}

\item{m}{fuzziness exponent (only for type = fuzzy)}

\item{build}{additional build algorithm to choose initial medoids (only
relevant for type = "fuzzy". Default FALSE.)}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
list with information on cluster results (medoid and cluster
assignment)
}
\description{
Function to perform pam in a fixed or fuzzy way on a data sample
}
