% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_sample.R
\name{perform_sample_clustering}
\alias{perform_sample_clustering}
\title{Perform pam or vegclust clustering on a data sample}
\usage{
perform_sample_clustering(
  dist,
  data,
  clusters,
  type,
  metric,
  names,
  m = 2,
  build = FALSE,
  verbose = 1,
  ...
)
}
\arguments{
\item{dist}{Dissimilarity matrix}

\item{data}{Data sample}

\item{clusters}{Number of clusters}

\item{type}{Hard or fuzzy clustering}

\item{metric}{A character specifying a predefined dissimilarity metric (like
\code{"euclidean"} or \code{"manhattan"}) or a self-defined dissimilarity
function. Defaults to \code{"euclidean"}. Will be passed as argument
\code{method} to \code{\link[proxy]{dist}}, so check \code{?proxy::dist} for
full details.}

\item{names}{Vector of names for observations}

\item{m}{Fuzziness exponent (only for type = fuzzy)}

\item{build}{Additional build algorithm to choose initial medoids (only
relevant for type = "fuzzy". Default FALSE.)}

\item{verbose}{Can be set to integers between 0 and 2 to control the level of
detail of the printed diagnostic messages. Higher numbers lead to more
detailed messages. Defaults to 1.}

\item{...}{Additional arguments passed to the main clustering algorithm
(\code{\link{pam}} or \code{\link[vegclust]{vegclust}})}
}
\value{
List with information on cluster results (medoid and cluster
assignment)
}
\description{
Function to perform pam in a hard or fuzzy way on a data sample
}
