% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_claraclust.R
\name{plot.fuzzyclara}
\alias{plot.fuzzyclara}
\title{Visualization of clustering solution by variables}
\usage{
\method{plot}{fuzzyclara}(x, data, type = NULL, confidence_threshold = 0, na.omit = FALSE, ...)
}
\arguments{
\item{x}{an object of class claraclust}

\item{data}{data.frame used for clustering}

\item{type}{type of plot}

\item{confidence_threshold}{threshold for fuzzy clustering observations to
be plotted}

\item{na.omit}{Should missing values be excluded for plotting?}

\item{...}{further arguments for plot functions}
}
\value{
clustering plot
}
\description{
Function to provide graphical visualization of distribution
}
