% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{plot.fuzzyclara}
\alias{plot.fuzzyclara}
\title{Visualization of clustering solution by variables}
\usage{
\method{plot}{fuzzyclara}(
  x,
  data,
  type = NULL,
  variable = NULL,
  na.omit = FALSE,
  membership_threshold = 0,
  sample_percentage = 1,
  plot_membership_scores = FALSE,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{An object of class "fuzzyclara"}

\item{data}{data.frame or matrix used for clustering}

\item{type, variable}{Type of plot. One of \code{c("barplot","boxplot","wordclouds",
"silhouette","pca","scatterplot", "parallel")}. Defaults to NULL, which either plots
a barplot or a boxplot, depending on the class of \code{variable}.}

\item{na.omit}{Should missing values be excluded for plotting? Defaults to
FALSE.}

\item{membership_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}

\item{sample_percentage}{Percentage value that indicates which percentage of
observations should randomly selected for representation the plot. Must be a
number between 0 and 1. Defaults to 1.}

\item{plot_membership_scores}{Boolean value indicating whether the cluster
membership scores for the observations should be indicated through line the
transparency (TRUE) or not (FALSE). Defaults to FALSE.}

\item{seed}{random number seed (needed for \code{clara_wordcloud} and
\code{clara_parallel})}

\item{...}{Further arguments for internal plot functions. For each type
there is an internal plot function. See for example \code{?clara_pca}.}
}
\value{
Clustering plot
}
\description{
Function to provide graphical visualization of distribution
}
\examples{

# Prepare data for example (enrich the USArrest dataset by area and state)
library(dplyr)
USArrests_enriched <- USArrests \%>\% 
  mutate(State = as.factor(rownames(USArrests)),
         Area  = as.factor(case_when(State \%in\% c("Washington", "Oregon",
                                               "California", "Nevada",
                                               "Arizona", "Idaho", "Montana",
                                               "Wyoming", "Colorado",
                                               "New Mexico", "Utah", "Hawaii",
                                               "Alaska") ~ "West",
                                  State \%in\% c("Texas", "Oklahoma", "Arkansas",
                                               "Louisiana", "Mississippi",
                                               "Alabama", "Tennessee",
                                               "Kentucky", "Georgia",
                                               "Florida", "South Carolina",
                                               "North Carolina", "Virginia",
                                               "West Virginia") ~ "South",
                                  State \%in\% c("Kansas", "Nebraska", "South Dakota",
                                               "North Dakota", "Minnesota",
                                               "Missouri", "Iowa", "Illinois",
                                               "Indiana", "Michigan", "Wisconsin",
                                               "Ohio") ~ "Midwest",
                                  State \%in\% c("Maine", "New Hampshire", "New York",
                                               "Massachusetts", "Rhode Island",
                                               "Vermont", "Pennsylvania",
                                               "New Jersey", "Connecticut",
                                               "Delaware", "Maryland") ~
                                               "Northeast")))
# Determine clusters that will be plotted                                 
cc_hard <- fuzzyclara(data        = USArrests,
                      clusters    = 3,
                      metric      = "euclidean",
                      samples     = 1,
                      sample_size = NULL,
                      type        = "hard",
                      seed        = 3526,
                      verbose     = 0)
cc_hard

cc_fuzzy <- fuzzyclara(data        = USArrests,
                      clusters    = 3,
                      metric      = "euclidean",
                      samples     = 1,
                      sample_size = NULL,
                      type        = "fuzzy",
                      m           = 2,
                      seed        = 3526,
                      verbose     = 0)
cc_fuzzy
                                    
# Boxplot
plot(x = cc_hard, data = USArrests_enriched, variable = "Assault") 

# Barplot
plot(x = cc_hard, data = USArrests_enriched, variable = "Area")

# Wordcloud
plot(x = cc_hard, data = USArrests_enriched, variable = "State", 
     type = "wordclouds", seed = 123)                                    
                                    
# Scatterplot
plot(x = cc_hard, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault")   

# Plot membership probability for fuzzy clustering
plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault", 
     focus = TRUE)

# Plot membership probability for fuzzy clustering (one cluster only)
plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault", 
     focus = TRUE, focus_clusters = c(1))  
     
# PCA
plot(x = cc_hard, data = USArrests_enriched, type = "pca",
     group_by = "Area")     
 
# Plot membership probability for one or more clusters following a PCA    
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     focus = TRUE)     
       
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     focus = TRUE, focus_clusters = c(1))      
  
# Silhouette plot
plot(x = cc_hard, data = USArrests, type = "silhouette")  
 
# Plot clusters for fuzzy clustering (using threshold for membership scores)
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     variable = "Assault", membership_threshold = 0) 
       
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     variable = "Assault", membership_threshold = 0.5)       
       
plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault", membership_threshold = 0)  
     
plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault", membership_threshold = 0.5)     
  
plot(x = cc_fuzzy, data = USArrests_enriched, type = "scatterplot",
     x_var = "Murder", y_var = "Assault", membership_threshold = 0.5,
     plot_all_fuzzy = TRUE)  
  
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     group_by = "Area", membership_threshold = 0)      
     
plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     group_by = "Area", membership_threshold = 0.5)     

plot(x = cc_fuzzy, data = USArrests_enriched, type = "pca",
     group_by = "Area", membership_threshold = 0.5, plot_all_fuzzy = TRUE)     

# Parallel Plot
plot(x = cc_fuzzy, data = USArrests_enriched, 
     type = "parallel", sample_percentage = 1, plot_membership_scores = TRUE)      
     
}
