% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzyclara.R
\name{plot.fuzzyclara}
\alias{plot.fuzzyclara}
\title{Visualization of clustering solution by variables}
\usage{
\method{plot}{fuzzyclara}(x, data, type = "boxplot", confidence_threshold = 0, na.omit = FALSE, ...)
}
\arguments{
\item{x}{An object of class claraclust}

\item{data}{data.frame used for clustering}

\item{type}{Type of plot. One of \code{c("barplot","boxplot","wordclouds",
"silhouette","pca","scatterplot")}. Defaults to NULL, which either plots
a barplot or a boxplot, depending on the class of \code{variable} specified
as part of the \code{...} argument.}

\item{confidence_threshold}{Threshold for fuzzy clustering observations to
be plotted. Must be a number between 0 and 1. Defaults to 0.}

\item{na.omit}{Should missing values be excluded for plotting? Defaults to
FALSE.}

\item{...}{Further arguments for internal plot functions.}
}
\value{
Clustering plot
}
\description{
Function to provide graphical visualization of distribution
}
