% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cluster_numbers.R
\name{plot_cluster_numbers}
\alias{plot_cluster_numbers}
\title{Function to provide graphical visualization for selecting the optimal number
of clusters.
The function provides graphical visualization showing the minimal (weighted)
average distance for every cluster number.}
\usage{
plot_cluster_numbers(cluster_results, clusters_range = NULL)
}
\arguments{
\item{cluster_results}{List of fuzzyclara clustering results with different
numbers of clusters.}

\item{clusters_range}{Optional range for the number of clusters. Defaults to
\code{NULL}.}
}
\description{
Function to provide graphical visualization for selecting the optimal number
of clusters.
The function provides graphical visualization showing the minimal (weighted)
average distance for every cluster number.
}
