% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fuzzyclara.R
\name{predict.fuzzyclara}
\alias{predict.fuzzyclara}
\title{Prediction of cluster assignments}
\usage{
\method{predict}{fuzzyclara}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class "fuzzyclara"}

\item{newdata}{data.frame to look variables with which to predict}

\item{...}{further arguments for predict functions}
}
\value{
clustering plot tibble
}
\description{
Function to predict cluster assignments
}
\examples{

# Split data intp test and training data
train_indices <- sample(x = nrow(USArrests), size = 0.7*nrow(USArrests))
USArrests_train <- USArrests[train_indices, ]
USArrests_test <- USArrests[-train_indices, ]

# Determine fuzzy clusters on training data

cc_fuzzy <- fuzzyclara(data        = USArrests_train,
                       clusters    = 3,
                       metric      = "euclidean",
                       samples     = 1,
                       sample_size = NULL,
                       type        = "fuzzy",
                       m           = 2,
                       seed        = 3526,
                       verbose     = 0)
cc_fuzzy

# Determine distance matrix of the observations and cluster medoids
USArrests_medoids <- USArrests[rownames(USArrests) \%in\% cc_fuzzy$medoids,]
dist <- proxy::dist(x = USArrests_test[, -1], y = USArrests_medoids[, -1],
                    method = "euclidean") #use same metric as above

# Make cluster prediction for test data

USArrests_clusters_predicted <- predict(object = cc_fuzzy,
                                        newdata = USArrests_test,
                                        dist_matrix = dist)
USArrests_clusters_predicted$membership_scores

}
